Added dataProvider module
